+++ spec/javascripts/govuk_publishing_components/analytics/external-link-tracker.spec.js
@@ -9,10 +9,10 @@ describe('GOVUK.analyticsPlugins.externalLinkTracker', function () {
   beforeEach(function () {
     $links = $(
       '<div class="external-links">' +
-        '<a href="http://www.nationalarchives.gov.uk"> National Archives </a>' +
-        '<a href="https://www.nationalarchives.gov.uk"></a>' +
-        '<a href="https://www.nationalarchives.gov.uk/one.pdf">National Archives PDF</a>' +
-        '<a href="https://www.nationalarchives.gov.uk/an/image/link.png"><img src="/img" /></a>' +
+        '<a href="http://www.nationalarchives1.gov.uk"> National Archives </a>' +
+        '<a href="https://www.nationalarchives2.gov.uk"></a>' +
+        '<a href="https://www.nationalarchives3.gov.uk/one.pdf">National Archives PDF</a>' +
+        '<a href="https://www.nationalarchives4.gov.uk/an/image/link.png"><img src="/img" /></a>' +
       '</div>' +
       '<div class="internal-links">' +
         '<a href="/some-path">Local link</a>' +
@@ -45,7 +45,7 @@ describe('GOVUK.analyticsPlugins.externalLinkTracker', function () {
     GOVUK.analyticsPlugins.externalLinkTracker({ externalLinkUploadCustomDimension: 36 })
 
     $('.external-links a').each(function () {
-      $(this).trigger('click')
+      GOVUK.triggerEvent($(this)[0], 'click')
       expect(GOVUK.analytics.trackEvent).toHaveBeenCalled()
       if (GOVUK.analytics.trackEvent.calls) {
         GOVUK.analytics.trackEvent.calls.reset()
@@ -53,7 +53,7 @@ describe('GOVUK.analyticsPlugins.externalLinkTracker', function () {
     })
 
     $('.internal-links a').each(function () {
-      $(this).trigger('click')
+      GOVUK.triggerEvent($(this)[0], 'click')
       expect(GOVUK.analytics.trackEvent).not.toHaveBeenCalled()
       if (GOVUK.analytics.trackEvent.calls) {
         GOVUK.analytics.trackEvent.calls.reset()
@@ -63,40 +63,52 @@ describe('GOVUK.analyticsPlugins.externalLinkTracker', function () {
 
   it('listens to click events on elements within external links', function () {
     GOVUK.analyticsPlugins.externalLinkTracker({ externalLinkUploadCustomDimension: 36 })
-    $('.external-links a img').trigger('click')
+    GOVUK.triggerEvent($('.external-links a img')[0], 'click')
 
     expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith(
-      'External Link Clicked', 'https://www.nationalarchives.gov.uk/an/image/link.png', { transport: 'beacon' })
+      'External Link Clicked', 'https://www.nationalarchives4.gov.uk/an/image/link.png', { transport: 'beacon' })
   })
 
   it('tracks an external link\'s href and link text', function () {
     GOVUK.analyticsPlugins.externalLinkTracker({ externalLinkUploadCustomDimension: 36 })
-    $('.external-links a').trigger('click')
+    var links = document.querySelectorAll('.external-links a')
+    for (var i = 0; i < links.length; i++) {
+      GOVUK.triggerEvent(links[i], 'click')
+    }
 
     expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith(
-      'External Link Clicked', 'http://www.nationalarchives.gov.uk', { transport: 'beacon', label: 'National Archives' })
+      'External Link Clicked', 'http://www.nationalarchives1.gov.uk', { transport: 'beacon', label: 'National Archives' })
 
     expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith(
-      'External Link Clicked', 'https://www.nationalarchives.gov.uk', { transport: 'beacon' })
+      'External Link Clicked', 'https://www.nationalarchives2.gov.uk', { transport: 'beacon' })
 
     expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith(
-      'External Link Clicked', 'https://www.nationalarchives.gov.uk/one.pdf', { transport: 'beacon', label: 'National Archives PDF' })
+      'External Link Clicked', 'https://www.nationalarchives3.gov.uk/one.pdf', { transport: 'beacon', label: 'National Archives PDF' })
+
+    expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith(
+      'External Link Clicked', 'https://www.nationalarchives4.gov.uk/an/image/link.png', { transport: 'beacon' })
   })
 
   it('duplicates the url info in a custom dimension to be used to join with a Google Analytics upload', function () {
     GOVUK.analyticsPlugins.externalLinkTracker({ externalLinkUploadCustomDimension: 36 })
-    $('.external-links a').trigger('click')
+    var links = document.querySelectorAll('.external-links a')
+    for (var i = 0; i < links.length; i++) {
+      GOVUK.triggerEvent(links[i], 'click')
+    }
 
-    expect(GOVUK.analytics.setDimension).toHaveBeenCalledWith(36, 'http://www.nationalarchives.gov.uk')
+    expect(GOVUK.analytics.setDimension).toHaveBeenCalledWith(36, 'http://www.nationalarchives1.gov.uk')
     expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith(
-      'External Link Clicked', 'http://www.nationalarchives.gov.uk', { transport: 'beacon', label: 'National Archives' })
+      'External Link Clicked', 'http://www.nationalarchives1.gov.uk', { transport: 'beacon', label: 'National Archives' })
   })
 
   it('does not duplicate the url info if a custom dimension is not provided', function () {
     GOVUK.analyticsPlugins.externalLinkTracker()
-    $('.external-links a').trigger('click')
+    var links = document.querySelectorAll('.external-links a')
+    for (var i = 0; i < links.length; i++) {
+      GOVUK.triggerEvent(links[i], 'click')
+    }
 
     expect(GOVUK.analytics.setDimension).not.toHaveBeenCalled()
-    expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith('External Link Clicked', 'http://www.nationalarchives.gov.uk', { transport: 'beacon', label: 'National Archives' })
+    expect(GOVUK.analytics.trackEvent).toHaveBeenCalledWith('External Link Clicked', 'http://www.nationalarchives1.gov.uk', { transport: 'beacon', label: 'National Archives' })
   })
 })